Lots of styling love for the new ticket form. It is probably the best form I've built yet.
Allow the user to select a business section when creating a ticket and save that as the component (list of available components goes in app_config.rb).

Remember the user and component in a cookie after they enter them in the new ticket form the first time.

Use cookies to keep 'flash' messages to show whether ticket posting was successful or not.

Rescue from the right exception when creating a new ticket.

Getting there!